+++ src/services/ui/src/components/Webform/index.tsx
@@ -1,7 +1,8 @@
+/* eslint-disable react/prop-types */
 import { useForm } from "react-hook-form";
 import { Button, Form } from "@/components/Inputs";
 import { RHFDocument } from "@/components/RHF";
-import { Alert, SubNavHeader } from "@/components";
+import { SubNavHeader } from "@/components";
 import { documentInitializer, documentValidator } from "@/components/RHF/utils";
 import { useGetForm } from "@/api";
 import { LoadingSpinner } from "@/components";
@@ -12,6 +13,7 @@ import { useLDClient } from "launchdarkly-react-client-sdk";
 import { featureFlags } from "shared-utils";
 import { useState } from "react";
 import { FormSchema } from "shared-types";
+import { MaintenanceBanner } from "@/components/MaintenanceBanner";
 
 export const Webforms = () => {
   return (
@@ -21,7 +23,7 @@ export const Webforms = () => {
       </SubNavHeader>
       <section className="block md:flex md:flex-row max-w-screen-xl m-auto px-4 lg:px-8 pt-8 gap-10">
         <div className="flex-1 space-x-5">
-          <UnderMaintenanceBanner />
+          <MaintenanceBanner />
           <Link
             path="/webform/:id/:version"
             params={{ id: "abp1", version: 202401 }}
@@ -79,14 +81,6 @@ function WebformBody({
   });
   const [subData, setSubData] = useState("");
 
-  const ldClient = useLDClient();
-  const clearDataButton: string = ldClient?.variation(
-    featureFlags.CLEAR_DATA_BUTTON.flag,
-    featureFlags.CLEAR_DATA_BUTTON.defaultValue
-  );
-
-  console.log(clearDataButton);
-
   const onSave = () => {
     const values = form.getValues();
     localStorage.setItem(`${id}v${version}`, JSON.stringify(values));
@@ -131,16 +125,7 @@ function WebformBody({
                   <Button type="button" onClick={onSave} variant="ghost">
                     Save draft
                   </Button>
-                  {clearDataButton && (
-                    <Button
-                      type="button"
-                      onClick={reset}
-                      variant="outline"
-                      className="mx-2"
-                    >
-                      Clear Data
-                    </Button>
-                  )}
+                  <ClearDataButton reset={reset} />
                 </div>
                 <Button type="submit">Submit</Button>
               </div>
@@ -182,35 +167,18 @@ export function Webform() {
   );
 }
 
-const UnderMaintenanceBanner = () => {
-  const banners = {
-    UNSCHEDULED: (
-      <h1 className="text-xl font-medium">Unschedule Maintenance Flag</h1>
-    ),
-    SCHEDULED: (
-      <h1 className="text-xl font-medium">Scheduled Maintenance Flag</h1>
-    ),
-  };
-
-  function getMaintenanceBanner(flag: string) {
-    return banners[flag as keyof typeof banners] || undefined;
-  }
-
+const ClearDataButton: React.FC<{ reset: () => void }> = ({ reset }) => {
   const ldClient = useLDClient();
-  const siteUnderMaintenanceBannerFlag = ldClient?.variation(
-    featureFlags.SITE_UNDER_MAINTENANCE_BANNER.flag,
-    featureFlags.SITE_UNDER_MAINTENANCE_BANNER.defaultValue
+  const clearDataButton: string = ldClient?.variation(
+    featureFlags.CLEAR_DATA_BUTTON.flag,
+    featureFlags.CLEAR_DATA_BUTTON.defaultValue
   );
 
-  const possibleMaintenanceBanner = getMaintenanceBanner(
-    siteUnderMaintenanceBannerFlag
-  );
-
-  if (possibleMaintenanceBanner) {
+  if (clearDataButton) {
     return (
-      <Alert className="mb-6 w-5/6" variant="destructive">
-        {possibleMaintenanceBanner}
-      </Alert>
+      <Button type="button" onClick={reset} variant="outline" className="mx-2">
+        Clear Data
+      </Button>
     );
   }
   return null;
