+++ src/services/ui/src/components/index.tsx
@@ -7,13 +7,14 @@ export * from "./Cards";
 export * from "./Chip";
 export * from "./Container";
 export * from "./DetailsSection";
+export * from "./Dialog";
 export * from "./ErrorAlert";
 export * from "./HowItWorks";
 export * from "./Layout";
 export * from "./LoadingSpinner";
+export * from "./MaintenanceBanner";
+export * from "./Modal";
+export * from "./Modal/ConfirmationModal";
 export * from "./PackageDetails";
 export * from "./SearchForm";
-export * from "./Modal";
-export * from "./Dialog";
-export * from "./Modal/ConfirmationModal";
 export * from "./Webform";
